[change]AutoSizeableTextを共通で使えるようにcommonファイルへ移動
